+++ web/src/pages/Settings/Streams/TypeSelectModal/TypeSelectModal.jsx
@@ -6,27 +6,33 @@ import Modal from "@/components/Modal/Modal";
 import Card from "@/components/Card/Card";
 import { getIcon } from "@/utils/getIcon";
 import ListItem from "@/components/List/ListItem/ListItem";
+import { useStatusStore } from "@/App";
 
 const TypeSelectModal = ({ onClose, onSelect }) => {
+    const stream_types_available = useStatusStore((s) => s.status.info.stream_types_available);
+
     return (
         <>
             <Modal className="streams-modal" onClose={onClose}>
                 <Card className="type-select-card">
                     <div className="type-select-title">Select A Stream Type</div>
                     <div>
-                        {StreamTemplates.filter((t) => !t.noCreate).map((t) => {
-                            return (
-                                <ListItem
-                                    key={t.type}
-                                    name={t.name}
-                                    onClick={() => {
-                                        onSelect(t.type);
-                                        onClose();
-                                    }}
-                                >
-                                    <img className="type-icon" src={getIcon(t.type)} />
-                                </ListItem>
-                            );
+                        {StreamTemplates
+                            .filter((t) => !t.noCreate)
+                            .filter((t) => stream_types_available.includes(t.type))
+                            .map((t) => {
+                                return (
+                                    <ListItem
+                                        key={t.type}
+                                        name={t.name}
+                                        onClick={() => {
+                                            onSelect(t.type);
+                                            onClose();
+                                        }}
+                                    >
+                                        <img className="type-icon" src={getIcon(t.type)} />
+                                    </ListItem>
+                                );
                         })}
                     </div>
                 </Card>